// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const LlmProvider = {
  GOOGLE: 'GOOGLE',
  OPENROUTER: 'OPENROUTER'
} as const

export type LlmProvider = (typeof LlmProvider)[keyof typeof LlmProvider]


export const ProjectRole = {
  OWNER: 'OWNER',
  EDITOR: 'EDITOR',
  VIEWER: 'VIEWER'
} as const

export type ProjectRole = (typeof ProjectRole)[keyof typeof ProjectRole]


export const DesignViewMode = {
  DESKTOP: 'DESKTOP',
  TABLET: 'TABLET',
  MOBILE: 'MOBILE'
} as const

export type DesignViewMode = (typeof DesignViewMode)[keyof typeof DesignViewMode]


export const ConnectionPosition = {
  TOP: 'TOP',
  RIGHT: 'RIGHT',
  BOTTOM: 'BOTTOM',
  LEFT: 'LEFT'
} as const

export type ConnectionPosition = (typeof ConnectionPosition)[keyof typeof ConnectionPosition]